# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Restaurant API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MenuItem {
  id: number;
  name: string;
  nameAr: string;
  category: string;
  price: number;
  description: string;
  descriptionAr: string;
  imageUrl: string;
  available: boolean;
  createdAt: string;
}

export interface CreateMenuItemRequest {
  name: string;
  nameAr: string;
  category: string;
  price: number;
  description: string;
  descriptionAr: string;
  imageUrl: string;
  available?: boolean;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export interface LoginResponse {
  success: boolean;
  message: string;
}

export interface AuthStatusResponse {
  authenticated: boolean;
  username?: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export type GetMenuItemsParams = {
  /**
   * Filter by category
   */
  category?: string;
  /**
   * Search by name or description
   */
  search?: string;
};