# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Restaurant API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get all menu items
 */
export const GetMenuItemsQueryParams = zod.object({
  category: zod.coerce.string().optional().describe("Filter by category"),
  search: zod.coerce
    .string()
    .optional()
    .describe("Search by name or description"),
});

export const GetMenuItemsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  nameAr: zod.string(),
  category: zod.string(),
  price: zod.number(),
  description: zod.string(),
  descriptionAr: zod.string(),
  imageUrl: zod.string(),
  available: zod.boolean(),
  createdAt: zod.coerce.date(),
});
export const GetMenuItemsResponse = zod.array(GetMenuItemsResponseItem);

/**
 * @summary Create a new menu item
 */
export const CreateMenuItemBody = zod.object({
  name: zod.string(),
  nameAr: zod.string(),
  category: zod.string(),
  price: zod.number(),
  description: zod.string(),
  descriptionAr: zod.string(),
  imageUrl: zod.string(),
  available: zod.boolean().optional(),
});

/**
 * @summary Update a menu item
 */
export const UpdateMenuItemParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateMenuItemBody = zod.object({
  name: zod.string(),
  nameAr: zod.string(),
  category: zod.string(),
  price: zod.number(),
  description: zod.string(),
  descriptionAr: zod.string(),
  imageUrl: zod.string(),
  available: zod.boolean().optional(),
});

export const UpdateMenuItemResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  nameAr: zod.string(),
  category: zod.string(),
  price: zod.number(),
  description: zod.string(),
  descriptionAr: zod.string(),
  imageUrl: zod.string(),
  available: zod.boolean(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete a menu item
 */
export const DeleteMenuItemParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteMenuItemResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});

/**
 * @summary Admin login
 */
export const LoginBody = zod.object({
  username: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});

/**
 * @summary Admin logout
 */
export const LogoutResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});

/**
 * @summary Get current auth status
 */
export const GetAuthMeResponse = zod.object({
  authenticated: zod.boolean(),
  username: zod.string().optional(),
});