# lib/api-zod/src/generated/types/authStatusResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Restaurant API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AuthStatusResponse {
  authenticated: boolean;
  username?: string;
}