# lib/api-zod/src/generated/types/createMenuItemRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Restaurant API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateMenuItemRequest {
  name: string;
  nameAr: string;
  category: string;
  price: number;
  description: string;
  descriptionAr: string;
  imageUrl: string;
  available?: boolean;
}